moving the webserver out of the library
