+++ pslink/Cargo.toml
@@ -62,6 +62,8 @@ actix-server = "1.0.4"
 tempdir = "0.3"
 test_bin = "0.3"
 tokio = "0.2.25"
+assert_cmd = "1.0.7"
+predicates = "2.0.0"
 
 [dev-dependencies.reqwest]
 features = ["cookies"]
